added crucible_vlan resource
Added a crucible_vlan resource to acquire and release VLANs from the Caster API.
See README for details.

- **NEW SETTING**: caster_api_url or SEI_CRUCIBLE_CASTER_API_URL as an env var. This
must be set to the url of the Caster API in order to use the crucible_vlan resource. Should
be set in the Caster API container to pass through the env var to the provider.
- **REQUIRED CONFIGURATION**: The oauth2/oidc client used by the provider in an
environment will need the proper scope(s) added to access the Caster API. Additionally,
the user used by the client will need to be a System Admin in Caster.
- all url settings are now normalized and will work with any combination of including or
not including trailing slashes or /api. The following examples will all properly resolve to
https://caster.example.com/api/ within the provider
  - https://caster.example.com/api/
  - https://caster.example.com/api
  - https://caster.example.com/
  - https://caster.example.com